Move sanity check into iterators module
