+++ src/hermite/mod.rs
@@ -105,26 +105,6 @@ mod tests {
     use approx::assert_abs_diff_eq;
 
     #[test]
-    fn iterator_sanity_check() {
-        for deg in (10..=100).step_by(10) {
-            let rule = GaussHermite::new(deg);
-            assert_eq!(rule.degree(), deg);
-            for ((ni, wi), (nn, ww)) in rule.iter().zip(rule.nodes().zip(rule.weights())) {
-                assert_abs_diff_eq!(ni, nn);
-                assert_eq!(wi, ww);
-            }
-            for ((ni, wi), (nn, ww)) in rule
-                .as_node_weight_pairs()
-                .iter()
-                .zip(rule.nodes().zip(rule.weights()))
-            {
-                assert_abs_diff_eq!(ni, nn);
-                assert_eq!(wi, ww);
-            }
-        }
-    }
-
-    #[test]
     fn golub_welsch_3() {
         let rule = GaussHermite::new(3);
         let x_should = [1.224_744_871_391_589, 0.0, -1.224_744_871_391_589];
